1. 실현손익 조회기능 추가 2. 매도시 주문가능수량 조회기능 추가 3. 알람메시지포맷팅 수정

+++ src/main/java/org/oopscraft/fintics/api/v1/dto/TradeAssetRequest.java
@@ -21,6 +21,6 @@ public class TradeAssetRequest {
 
     private boolean enabled;
 
-    private Double holdRatio;
+    private BigDecimal holdRatio;
 
 }
