+++ src/main/java/org/oopscraft/fintics/model/OrderBook.java
@@ -4,16 +4,18 @@ import lombok.*;
 import lombok.experimental.SuperBuilder;
 import net.bytebuddy.implementation.bind.annotation.Super;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 public class OrderBook {
 
-    Double price;
+    BigDecimal price;
 
-    Double bidPrice;
+    BigDecimal bidPrice;
 
-    Double askPrice;
+    BigDecimal askPrice;
 
 }
